requirements.txt
-i https://pypi.org/simple
blinker==1.7.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
dnspython==2.4.2; python_version >= '3.8' and python_version < '4.0'
email-validator==2.1.0.post1; python_version >= '3.8'
flask==3.0.0; python_version >= '3.8'
flask-login==0.6.3; python_version >= '3.7'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
flask-wtf==1.2.1; python_version >= '3.8'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.6; python_version >= '3.5'
importlib-metadata==7.0.1; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
sqlalchemy==2.0.23; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
werkzeug==3.0.1; python_version >= '3.8'
wtforms==3.1.1; python_version >= '3.8'
zipp==3.17.0; python_version >= '3.8'
